trash

+++ ParallelMPISF/AgentOnNodeInfo.h
@@ -4,8 +4,8 @@ class AgentOnNodeInfo
 public:
 	AgentOnNodeInfo(void): isDeleted(false), _nodeID(-1), _agentID(-1) { };
 	bool isDeleted;
-	size_t _nodeID;
-	size_t _agentID;
+	int _nodeID;
+	long long _agentID;
 	AgentOnNodeInfo(int nodeID, int agentID) : isDeleted(false), _nodeID(nodeID), _agentID(agentID) { }
 	AgentOnNodeInfo(int nodeID, int agentID, bool isDeleted) : isDeleted(isDeleted), _nodeID(nodeID), _agentID(agentID) { }
 	~AgentOnNodeInfo(void);
